update names of classes, separate components to be imported under 'Picker'
